requirements.txt
yfinance>=0.2.36
pandas>=1.5.0
requests>=2.28.0
schedule>=1.2.0
python-dotenv>=0.21.0
tradingview-ta
pytz